Fix menu box headers
In some wikis, the menu box headers were displayed as "<Navigazione>"
instead of just "Navigazione".

This was caused by the translation magic for the box header, which I
consider superfluous - Mediawiki:Sidebar in the wikis I checked already
contains the localized headers.

Removing the translation magic fixes the problem with the superfluous
angle brackets.